lib: Import getdate source from gnulib.
Each of these files was copied directly from the gnulib source
repository at git commit ID e961f4c04b6bbd1bf6f98c91fa0e6ae1d7a8eea9
The git repository was cloned from:

	git://git.sv.gnu.org/gnulib.git

Each file here is made available either under the GNU GPLv2+ or the
GNU GPLv3+ so all are compatible with the license of notmuch (GNU
GPLv3+).

The import was made with the help of the import script in this
directory, which can be run again in the future to import new changes
to the gnulib source.